+++ DEMU_ANDROID/app/src/main/java/com/example/demu_android/SplashActivity.kt
@@ -7,8 +7,6 @@ import android.os.Handler
 import android.os.Looper
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
 
 class SplashActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -20,7 +18,7 @@ class SplashActivity : AppCompatActivity() {
     }
 
     private fun moveToMainActivity() {
-        val intent = Intent(this, MainActivity::class.java)
+        val intent = Intent(this, Main2Activity::class.java)
 
         Handler(Looper.getMainLooper()).postDelayed( {
             startActivity(intent)
