+++ DEMU_ANDROID/app/src/main/res/menu/menu_bottom_view.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/fragment_home"
+        android:title="홈"
+        android:icon="@drawable/ic_home" />
+    <item
+        android:id="@+id/fragment_search"
+        android:title="검색"
+        android:icon="@drawable/ic_search" />
+    <item
+        android:id="@+id/fragment_blog"
+        android:title="블로그"
+        android:icon="@drawable/ic_write_blog" />
+    <item
+        android:id="@+id/fragment_my_page"
+        android:title="마이페이지"
+        android:icon="@drawable/ic_person"
+        />
+</menu>
