+++ DEMU_ANDROID/app/src/main/res/values/menu.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+
+</resources>
